Finish migrating server state
