--- src/serverState/_DEPRECATED.md
@@ -1,2 +1,0 @@
-Using react-query directly in this project is deprecated. Use `jotai-tanstack-query` in
-the `state/server/` directory.
